STYLE: vtkMRMLNode: Fix "sign-compare" warning
Fix "sign-compare" warning originally introduced in r25028 (ENH: Added
Python-wrappable method for getting node references)

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=25070
git-svn-id: http://svn.slicer.org/Slicer4/trunk@25070 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Core/vtkMRMLNode.cxx
@@ -1056,9 +1056,9 @@ int vtkMRMLNode::GetNumberOfNodeReferenceRoles()
 //----------------------------------------------------------------------------
 const char* vtkMRMLNode::GetNthNodeReferenceRole(int n)
 {
-  if (n<0 || n>=this->NodeReferences.size())
+  if (n < 0 || n >= static_cast<int>(this->NodeReferences.size()))
     {
-    vtkErrorMacro("vtkMRMLNode::GetNthNodeReferenceRole failed: n="<<n<<" is out of range");
+    vtkErrorMacro("vtkMRMLNode::GetNthNodeReferenceRole failed: n=" << n << " is out of range");
     return NULL;
     }
   NodeReferencesType::iterator roleIt( this->NodeReferences.begin() );
